requirements.txt
streamlit
google-generativeai
google-genai
python-dotenv
speechrecognition
pyaudio
reportlab
